Get some *Help*, *Apropos*, and *Man* maps back
In trying to get Evil conformity, the mappings for these modes are
getting overwritten.

There are two ways around this:

One is to figure out a keymap to have in these buffers that match each
other as well as the default Emacs ones do and apply this in Evil mode.

Another is to keep the Emacs bindings as much as possible, and remove
Evil conflicts wherever they get in the way.

This commit is the first step in the second method, this one has only
been chosen as it's the easiest to get started with -- and I can apply
fixes whenever something bothers me without worrying about the
implications.