Initial, dumb method to swap keys e.g. 9 and (
In vim I have all the number keys mapped to their shifted equivalent
when in insert mode.
I want similar in emacs, mainly when writing lisp I don't want to be
pressing shift all the time for the braces.

This commit makes the first steps towards this, but it's not pretty,
most of the clever ways I can think of writing this aren't
working (usually because of confusion with lexical bindings in emacs)