Emacs `transpose-subr' has changed
I had an advice written around Emacs's `transpose-subr' function in
order to change its functionality.
It turns out that the functionality I wanted was what it used to be, and
the alternate functionality was a regression.
Newer versions of Emacs have fixed this regression so I don't need that
advice.

As I no longer change the normal operation of this Emacs function I can
use the move-text package instead of keeping a modified version in my
own config.