Define `query-search-and-replace-multi' command
This does similar to the :S/{,}/{,} command from vim-abolish.
It takes a list of (match replacement) pairs, and replaces each occurance of the
match with its corresponding replacement.

At the moment I have it ask for confirmation as this is the default for M-% and
C-M-% bindings, so it fits in with what I expect when I'm using it.

There is no option to run without confirmation, this could easily be added,
we'll see if I want it in the future.