Remove customisation for the `python-shell'
Recent updates to ipython mean my customisations stopped working and I got a
load of control characters around the prompt.
I gave a little effort to fixing it, but it's much easier to just use the
default options and the default python version.

Anyway, the only feature I use of IPython is the ability to reference previous
outputs easily (via Out[<number>]), and I can get by without that.