Use the new cask--initialize function
It seems that cask is secifically moving away from what I wanted to use
it for.  I want to use it to download all my packages for emacs, and
then set up the include path to point to the packages that it has
downloaded.

It seems that cask has started to put this functionality in private
functions rather than public ones.

At some point in the future I will need to do the proper inspection and
figure out whether cask is still the way to go or if there's some other
approach that would be better.  For now I'll just use the private
function and carry on.