Do not setup windmove keybindings
I'm not using these at the moment, and with an update to emacs it seems
that these are now conflicting with org-mode keybindings.

Rather than try to control the ordering of the keymaps in order to mask
out the windmove keybindings, it seems neater to not set up the windmove
keybindings in the first place.