Adjustments for yas--auto-fill strangeness
yasnippet has some problems around auto-fill.  At some point the
"saved" auto-fill function is lost (a known bug upstream but it's not
known why).  When this happens `yas--auto-fill' turns `auto-fill-mode'
off.

The consistent turning of `auto-fill-mode' off is inconvenient for me,
especially since I know what `yas--original-auto-fill-function' should
be and can reset it directly.  Here I overwrite `yas--auto-fill' to
reset `yas--original-auto-fill-function' instead of turning
`auto-fill-mode' off.

On top of that, in order to aid debugging, I add a variable-watcher to
alert me when this variable changes.  Hopefully I will notice the
change at the time it happens and be able to associate it with some
action I was performing at the time.