Ensure text coming from process are "undo joined" with the rest
There is the concept of an `undo-boundary' in emacs.  This indicates
how many changes a single call to `undo' would remove.

When an underlying process is outputting text slowly into a VSH
buffer, I want all that text to be in one `undo'.  Conversely, I don't
want any of the text to be in the same `undo' as a modification made
by the user elsewhere at the same time.

This commit satisfies both requirements by ensuring there are `undo'
boundaries after every insert of text come from the underlying
process, but then removing those boundaries when the next change is
also text coming from the underlying process.