Introduce relative-path commands based on underlying process cwd
Any terminal is going to have a foreground process group.  On Linux
systems we can find the foreground process group for the terminal of a
VSH buffer, find the working directory of that process group, and then
run some commands (like `find-file') as if from that directory.

This allows the relevant commands to provide hints based on the
surrounding text in the given buffer.