Change mappings for `vsh-save-command' and `vsh-activate-command'
The newly added keybinding for `vsh-find-file' uses `C-c C-x' as a
prefix, but I'd mapped that to `vsh-activate-command'.

I want `C-c C-x C-f' as the keybinding for `vsh-complete-file-name',
so I changed the bindings for `vsh-save-command' and
`vsh-activate-command' to use `C-c s' and `C-c x' intead of `C-c C-s'
and `C-c C-x' respectively.