Do not keyswap -/_ in erts-mode
This mode is for testing emacs lisp.  Some of the time one would be
testing lisp functions that are supposed to act on code written in
other languages.

However, all of the time one would be likely to write a `Code:'
specification in emacs lisp.  Plus some of the time I would be writing
tests to run on some kind of lisp.

I happen to mostly be using this mode for testing `vsh-mode' buffer
commands.  These modes do not need keyswap to help.  Hence choosing
the keyswap settings such that they help writing the lisp commands to
run for tests.