Properly return `proc' from `vsh-start-process'
When starting to run tests I hit crashes and found this invalid
setting.  Returning `proc' from outside the `let' statement that bound
it.

It didn't actually matter since I didn't use the return value
anywhere, but the fix is easy and I "feel like" this function should
return this value.