Random save-point in re-working regexp patterns
I've been wanting to neaten these up for a while, and this way I'm
getting to that.  The changes in definition involved here are:
- Comment now encompasses:
  - Any full prompt where the first non-blank character is a hash.
    (No longer needs the single space after the hash, no longer needs
    no spaces between prompt and hash).
  - prompt without it's trailing whitespace.

On top of this, I've adjusted the regexp patterns to make other
functions easier to implement -- match-string 1 represents "everything
up to whitespace before text" and match-string 2 represents
"whitespace before text".  This makes implementing
`vsh--line-beginning-position' much easier.