Move the vsh.el package into a separate repository
The repository is shared with vsh for vim.  As it stands I have to
run one shell command and one emacs command in order to install the
package -- will hopefully get this up onto MELPA so I can make the
installation fit with everything else.

For reference, the installation is done through:
 vshcmd: > cd ~/.emacs.d/vsh
 vshcmd: > ./emacs-tar.sh
Then inside emacs:
 (package-install-file "~/.emacs.d/vsh/vsh-1.0.tar")