Re-working init file
I had originally thought it would be good to use `cask` for
initialisation and in order to run tests with `ecukes`.

It later seems that the functionality I was using in `cask` is not
supported.  This meant I had to avoid updating `cask`, and that
updates were getting problematic.

Have now gone and found what I believe to be the recommended mechanism
for ensuring all packages are installed when starting emacs -- this
being `(package-install-selected-packages t)` and using the
automatically maintained list of installed packages in `customize`.

While doing this I took the opportunity to remove some packages that I
rarely use and think I could do better with.  These being `smart-tab`
and `yasnippet`.

`yasnippet` had some problems with `auto-fill-mode` and
even though I thought this was fixed it does seem there are still
problems.  I honestly don't know how often I use `smart-tab`.  We will
see how completion vs indentation feels for the moment.