Remove old flymake settings and increase timeout for eglot/flymake
Was finding the automatic updates of flymake and eglot to get in the
way of typing.  When pausing while inserting new text into the buffer
I would accidentally trigger a refresh and that would delay my typing.

Here we adjust the timeouts for sending changes over to the LSP server
and for updating the flymake errors.