Adjust default mappings for completion in buffer
I would like to use C-n, C-p, and RET, but they interfere with typing
too much.  I find that I've gotten the completion that I want first
time and want to start a new line but am stuck in this completion
mode.

Using M-n and M-p, these also interfere with some mappings just not as
many and I don't hit that problem often.