Adjust `goto-error-on-linenum' for more modes
As it stood this was only really working in compile buffers and grep
buffers.  I'd like it to work for more things (e.g. occur).

I think using "whatever is bound to RET" at that point is likely going
to be the most robust implementation so that's what I'm going with for
the moment.