Adjust fonts to use
I still prefer Tamsyn in general, but it doesn't have any display for
some of the unicode characters which is getting more annoying now that
some basic command line apps emit these to be pretty (one that
triggered the switch just now was `xinput list').

Going with something that I understand to be more supported, have
these unicode characters, and be good enough -- Liberation Mono.
N.b. one nice feature of this is that it shows up in `fc-list',
because IIUC it has a client-side font.